Update _index.md
Since "2020)" was on its own separate line, markdown interpreted it as a start of an ordered list. Just like this.

1)
2)
3)

And then when that markdown got converted to HTML, it got converted into a literal <ol> tag, which stands for Ordered List - http://html5doctor.com/element-index/#ol

However, in order for an Ordered List to show anything, it must have within it an <li> tag, which stands for List Item. Since there is no such tag, HTML treats this as an empty list and shows nothing, resulting in the bug.

Here is a snippet from the literal HTML from this link -- https://openjdk.org/projects/amber/

<li><a href="design-notes/patterns/type-patterns-in-switch">Type Patterns in <code>switch</code></a> (September
<ol type="1" start="2020"></ol>
</li>