Fixing the incorrect code example
There are only 3 methods on java.util.regex.Matcher that have the name group. Here they are.

https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/util/regex/Matcher.html#group()

https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/util/regex/Matcher.html#group(int)

https://docs.oracle.com/en/java/javase/21/docs/api/java.base/java/util/regex/Matcher.html#group(java.lang.String)

And since this is an IntStream, the only possible method that could be matched is the 2nd link above. Based on that, it is clear that the code example meant to put an `m` instead of `Matcher`.